% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_frontier.R
\name{get_frontier}
\alias{get_frontier}
\title{Get values on the efficient frontier}
\usage{
get_frontier(data, x, y, quadrant = c("top.right", "bottom.right",
  "bottom.left", "top.left"))
}
\arguments{
\item{data}{A data frame}

\item{x}{Numeric vector of values}

\item{y}{Numeric vector of values}

\item{quadrant}{Depending on which direction on each axes defines "better",
it places the frontier in the appropriate quadrant.}
}
\value{
A Kraljic purchasing matrix plot
}
\description{
\code{get_frontier} identifies and extracts the \code{x} and \code{y} values
that fall along the efficient frontier.
}
\examples{

# Given the following \\code{x} and \\code{y} attribute values we can plot each
# product or service in the purchasing matrix:
library(dplyr)

psc2 <- mutate(psc, x_SAVF_score = SAVF_score(x_attribute, 1, 5, .653),
                    y_SAVF_score = SAVF_score(y_attribute, 1, 10, .7))

kraljic_matrix(psc2, x_SAVF_score, y_SAVF_score)

}
