% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_frontier.R
\name{geom_frontier}
\alias{geom_frontier}
\title{Plot values on the efficient frontier with a ggplot2 object}
\usage{
geom_frontier(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", direction = "vh", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()} or
\code{aes()}. If specified and `inherit.aes = TRUE` (the
default), it is combined with the default mapping at the top level of the
plot. You must supply `mapping` if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If `NULL`, the default, the data is inherited from the plot
   data as specified in the call to `ggplot()`.

   A `data.frame`, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   [fortify()] for which variables will be created.

   A `function` will be called with a single argument,
   the plot data. The return value must be a `data.frame.`, and
   will be used as the layer data.}

\item{direction}{direction of stairs: 'vh' for vertical then horizontal, or
'hv' for horizontal then vertical}

\item{...}{other arguments passed on to layer. These are often aesthetics,
used to set an aesthetic to a fixed value, like `color = "red"` or `size = 3`.}

\item{x}{Numeric vector of values}

\item{y}{Numeric vector of values}

\item{quadrant}{Depending on which direction on each axes defines "better",
it places the frontier in the appropriate quadrant. For example, if large
\code{x} and large \code{y} values are considered "better" then the default
"top.right" is appropriate.  If small \code{x} and small \code{y} are
considered "better" then "bottom.left" is appropriate.}
}
\description{
\code{geom_frontier} plots the \code{x} and \code{y} values
that fall along the user defined efficient frontier. The efficient frontier
can be defined by a combination of large \code{x} and large \code{y} values,
large \code{x} and small \code{y} values, small \code{x} and large \code{y}
values, or small \code{x} and small \code{y} values.
}
\section{Aesthetics}{


`geom_frontier` understands the following aesthetics:
    \enumerate{
      \item \code{x}:
      \item \code{y}:
      \item \code{alpha}:
      \item \code{color}:
      \item \code{linetype}:
      \item \code{Size}:
    }
}

\examples{

}
\seealso{
\code{\link{geom_frontier}} for identifying and extracting the observations that
lie along the efficient frontier.
}
