% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frontier.R
\name{get_frontier}
\alias{get_frontier}
\title{Compute the Pareto Frontier}
\usage{
get_frontier(x, which.cols = 1:2, quadrant = c("top.right", "bottom.right",
  "bottom.left", "top.left"))
}
\arguments{
\item{x}{A data frame.}

\item{which.cols}{If \code{x} is a data frame, then which columns to use in
computing the efficient frontier.}

\item{quadrant}{Chararacter string specifying which quadrant the frontier
should appear in. Default is \code{"top.right"}.}
}
\value{
A data frame containing the data points that make up the efficient
frontier.
}
\description{
Extract the points that make up the pareto frontier from a set of data.
}
