% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frontier.R
\name{get_frontier}
\alias{get_frontier}
\alias{get_frontier.data.frame}
\alias{get_frontier.default}
\title{Compute the Pareto Frontier}
\usage{
get_frontier(x, y, quadrant = c("top.right", "bottom.right", "bottom.left",
  "top.left"))

\method{get_frontier}{data.frame}(x, y, quadrant = c("top.right",
  "bottom.right", "bottom.left", "top.left"))

\method{get_frontier}{default}(x, y, quadrant = c("top.right", "bottom.right",
  "bottom.left", "top.left"))
}
\arguments{
\item{x}{A data frame or numeric vector.}

\item{y}{A numeric vector.}

\item{quadrant}{Chararacter string specifying which quadrant the frontier
should appear in. Default is \code{"top.right"}.}
}
\value{
A data frame containing the data points that make up the efficient
frontier.
}
\description{
Extract the points that make up the pareto frontier from a set of data.
}
